requirements.txt
librosa
soundfile
google-genai
python-dotenv
